create megabot
