+++ BotModules/MemeNarc/config.json
