+++ README.md
@@ -1,9 +1,5 @@
 # Discord-Bots
 
-
-## MemeNarc
-Removes Memes that weren't put in a meme chat.
-
 ## Budgett
 Named after the "it's Wednsesday my dude" frog.
 
@@ -16,7 +12,13 @@ This bot
 #### Work to be added
 Add an array of new vidos for the bot to post or give the bot a way to find videos itself.
 
-## ScreechBot
+## Megabot
+This is a framework bot for holding other bot modules without having to run a seperate bot token for each of them. The following bots are submodules of megabot.
+
+### MemeNarc
+Removes Memes that weren't put in a meme chat.
+
+### ScreechBot
 Has several chat commands including /screech which will have the bot screech into the voice chat. Primarily hangs in voice chat and greets newcomers.
 
 Use /help to list available commands when running
